+++ CHANGES.md
@@ -3,6 +3,7 @@
 * add `gu_plugin_name()` to return plugin name, slug or slug-didhash
 * change `can_update()` check to `manage_options` for `DISALLOW_FILE_MODS` constant
 * MIT to GPL-3.0-or-later because of distributed components, etc
+* cast `$response` elements to object in `parse_contents_response()`
 
 #### 12.20.2 / 2025-12-08
 * harden REST API data for versions if relesase_assets and tags are empty -- this can happen if too many tags are created that aren't semver format